add startSubmission mutation and extract data to json files
