+++ src/mock-data/authenticatedUser.json
@@ -1,0 +1,13 @@
+{
+    "token_id": "fd7d795a-67bc-4266-a426-722170a06e7d",
+    "token_version": "0.1-alpha",
+    "identity": {
+        "user_id": "e4b91392-52d5-4c04-838a-9a5006d46d73",
+        "external": [
+            { "id": "ewwboc7m", "domain": "elife-profiles" },
+            { "id": "0000-0002-6277-0372", "domain": "orcid" }
+        ]
+    },
+    "roles": [{ "journal": "elife", "kind": "author" }],
+    "meta": null
+}
