+++ src/mock-data/journalUser.json
@@ -1,0 +1,5 @@
+{
+    "iss": "journal--prod",
+    "id": "ewwboc7m",
+    "new-session": true
+}
