+++ src/mock-graphql/resolvers.ts
@@ -1,7 +1,12 @@
-import { getSubmissions } from '../use-cases';
+import { getSubmissions, startSubmission } from '../use-cases';
+
+const submissions = [];
 
 export const resolvers = {
     Query: {
-        getSubmissions,
+        getSubmissions: getSubmissions(submissions),
+    },
+    Mutation: {
+        startSubmission: startSubmission(submissions),
     },
 };
