+++ src/use-cases/getSubmission.ts
@@ -1,0 +1,3 @@
+export const getSubmission = (submissions: {}): (() => Array<{}>) => () => {
+    return submissions[0];
+};
