+++ src/use-cases/getSubmissions.ts
@@ -1,5 +1,1 @@
-import submissionsData from '../data/submissions.json';
-
-export const getSubmissions = (): Array<{}> => {
-    return submissionsData;
-};
+export const getSubmissions = (submissions: Array<{}>): (() => Array<{}>) => () => submissions;
