+++ src/use-cases/index.ts
@@ -3,6 +3,7 @@ import { Request, Response } from 'express';
 export * from './journalSubmit';
 export * from './authenticate';
 export * from './getSubmissions';
+export * from './startSubmission';
 
 export const HealthCheck = () => (req: Request, res: Response): void => {
     res.json({ ok: true });
